refactor(task-lease): isolate local record normalization
